Mark as completed all requests, which come from the same ec2 instanance, ask for the same permissions set and was already successfully processed.
When there are multiple apps deployed on the same ec2,
which use aws-integration interface and all of them request for the same roles
and policies set, only the first request is successfully processed.
Every next request coming from the same ec2 instance is ignored,
because it is treated as already done. Ignored requests are never marked
as complted.